+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Improve performance of write_file for Docker sandboxes.
 - Use user_data_dir rather than user_runtime_dir for view notifications.
 - Implement `read_eval_log_sample()` for JSON log files.
+- Log the list of dataset sample IDs.
 
 ## v0.3.42 (23 October 2024)